requirements.txt
asgiref==3.7.2
attrs==25.1.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.5
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.2.0
django-database-url==1.0.3
django-extensions==3.2.3
django-filter==23.5
django-jazzmin==2.6.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-spectacular==0.28.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.23.1
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.7
